Relocated lib files to src
