+++ src/packedbuffer.js
